big cleanup for profile lookup logic
