+++ src/detail-panels/packs/pack-view.jsx
@@ -1,95 +1,99 @@
 // @ts-check
- 
+
 import { FormatTimestamp } from '../../common-components/format-timestamp';
-import {useResolveDidToProfile} from '../../api/resolve-handle-or-did'; 
-import "./list-packs.css"
+import { useResolveDidToProfile } from '../../api/resolve/did-to-profile';
+import './list-packs.css';
 import { ConditionalAnchor } from '../../common-components/conditional-anchor';
 
 /**
  * @param {{
-*  className?: string,
-*  packs?: PackListEntry[]
-* }}_
-*/
-export function PackView({packs, className=""}){ 
+ *  className?: string,
+ *  packs?: PackListEntry[]
+ * }}_
+ */
+export function PackView({ packs, className = '' }) {
   return (
-    <ul className={"packs-as-pack-view "+ (className || '')}>
-      {packs && packs?.length>0 &&  packs.map( (pack,i)=> 
-        <PackViewEntry key={i} entry={pack} />
-       )}
+    <ul className={'packs-as-pack-view ' + (className || '')}>
+      {packs &&
+        packs?.length > 0 &&
+        packs.map((pack, i) => <PackViewEntry key={i} entry={pack} />)}
     </ul>
-
   );
 
   /**
    * @param {{
-  *  className?: string,
-  *  entry: PackListEntry
-  * }} _
-  */
-  function PackViewEntry({className="", entry}){ 
-
+   *  className?: string,
+   *  entry: PackListEntry
+   * }} _
+   */
+  function PackViewEntry({ className = '', entry }) {
     const originator = useResolveDidToProfile(entry.did);
- 
-    return(
+
+    return (
       <li className={'pack-entry ' + (className || '')}>
-      <div className='row'>
-      
-       <span className='pack-name'>
-        <ConditionalAnchor 
-          condition={(!originator.isError && originator.data)}
-          href={entry.url??""} 
-          target='__blank'>
-          <AvatarAndName name={entry.name} avatarUrl={originator.data?.avatarUrl} />
-        </ConditionalAnchor>
-        </span>
-        <FormatTimestamp
-          timestamp={entry.created_date ?? ""}
-          noTooltip
-          className='pack-add-date' />
-      </div>
-      <div className='row'>      
-        
-        <span className='pack-description'>
-          {entry.description && ' ' + entry.description}
-        </span>
-      </div>
-    </li>
+        <div className="row">
+          <span className="pack-name">
+            <ConditionalAnchor
+              condition={!originator.isError && originator.data}
+              href={entry.url ?? ''}
+              target="__blank"
+            >
+              <AvatarAndName
+                name={entry.name}
+                avatarUrl={originator.data?.avatarUrl}
+              />
+            </ConditionalAnchor>
+          </span>
+          <FormatTimestamp
+            timestamp={entry.created_date ?? ''}
+            noTooltip
+            className="pack-add-date"
+          />
+        </div>
+        <div className="row">
+          <span className="pack-description">
+            {entry.description && ' ' + entry.description}
+          </span>
+        </div>
+      </li>
     );
   }
-/**
+  /**
    * @param {{
    *  avatarUrl?: string,
-    *  name: string
-    * }} _
-*/
-  function  AvatarAndName ({ avatarUrl, name}){
+   *  name: string
+   * }} _
+   */
+  function AvatarAndName({ avatarUrl, name }) {
     return (
       <>
-        <span 
-          className='pack-creator-avatar'
+        <span
+          className="pack-creator-avatar"
           style={
             !avatarUrl
-              ? { background:'none', 
-                  color:'cornflowerblue',
-                  textAlign:'center', 
-                  transform:'scale(1.5)'}
-              : { 
-                  backgroundPosition:'center',
-                  color:'transparent',
-                  borderRadius:"200%",
+              ? {
+                  background: 'none',
+                  color: 'cornflowerblue',
+                  textAlign: 'center',
+                  transform: 'scale(1.5)',
+                }
+              : {
+                  backgroundPosition: 'center',
+                  color: 'transparent',
+                  borderRadius: '200%',
                   backgroundImage: `url(${avatarUrl})`,
                   animationDelay: '10s',
                 }
           }
-        >ⓓ
+        >
+          ⓓ
         </span>
         {name}
       </>
     );
   }
 
-    /*
+  /*
 
     if (isLoading) {
         return (
@@ -111,5 +115,3 @@ export function PackView({packs, className=""}){
     );
     */
 }
-
-
